Cleanup, refactoring (FindBugs vs. SpotBugs annotations, warnings).
