+++ src/main/java/io/jenkins/plugins/analysis/core/model/LabelProviderFactory.java
@@ -1,16 +1,17 @@
 package io.jenkins.plugins.analysis.core.model;
 
-import javax.annotation.CheckForNull;
 import java.util.List;
+import javax.annotation.CheckForNull;
 
 import org.apache.commons.lang.StringUtils;
 
-import edu.hm.hafner.util.VisibleForTesting;
 import io.jenkins.plugins.analysis.core.JenkinsFacade;
 import io.jenkins.plugins.analysis.core.model.StaticAnalysisTool.StaticAnalysisToolDescriptor;
 
 import hudson.DescriptorExtensionList;
 import hudson.ExtensionPoint;
+
+import edu.hm.hafner.util.VisibleForTesting;
 
 /**
  * Creates {@link StaticAnalysisLabelProvider} instances based on a provided ID and name.
@@ -74,7 +75,7 @@ public class LabelProviderFactory {
             }
         }
 
-        return new DefaultLabelProvider(id, name);
+        return new StaticAnalysisLabelProvider(id, name);
     }
 
     private StaticAnalysisLabelProvider wrapLabelProvider(final StaticAnalysisLabelProvider labelProvider,
