+++ src/main/java/io/jenkins/plugins/analysis/core/model/Summary.java
@@ -74,95 +74,6 @@ public class Summary {
         return Messages.Tool_ParticipatingTools(tools);
     }
 
-    /*
-    @Override
-    public String toString() {
-        HtmlPrinter printer = new HtmlPrinter();
-        printer.append(createDeltaMessage());
-
-        if (result.getTotalSize() == 0 && result.getNoIssuesSinceBuild() > 0) {
-            printer.append(printer.item(Messages.ResultAction_NoWarningsSince(result.getNoIssuesSinceBuild())));
-            printer.append(printer.item(createHighScoreMessage()));
-        }
-        else if (result.isSuccessfulTouched()) {
-            printer.append(printer.item(createPluginResultMessage()));
-            if (result.isSuccessful()) {
-                printer.append(printer.item(createSuccessfulHighScoreMessage()));
-            }
-        }
-        return printer.toString();
-    }
-
-    private String createPluginResultMessage() {
-        return Messages.ResultAction_Status() + getResultIcon() + " - " + result.getReason() + getReferenceBuildUrl();
-    }
-
-    private String getReferenceBuildUrl() {
-        HtmlPrinter printer = new HtmlPrinter();
-        int referenceBuild = result.getReferenceBuild();
-        if (referenceBuild > 0) {
-            printer.append("&nbsp;");
-            printer.append("(");
-            printer.append(Messages.ReferenceBuild());
-            printer.append(": ");
-            printer.append(printer.link(String.format("../%d", referenceBuild), String.format("#%d", referenceBuild)));
-            printer.append(")");
-        }
-        return printer.toString();
-    }
-
-    private String createHighScoreMessage() {
-        if (result.isNewZeroWarningsHighScore()) {
-            long days = getDays(result.getZeroWarningsHighScore());
-            if (days == 1) {
-                return Messages.ResultAction_OneHighScore();
-            }
-            else {
-                return Messages.ResultAction_MultipleHighScore(days);
-            }
-        }
-        else {
-            long days = getDays(result.getHighScoreGap());
-            if (days == 1) {
-                return Messages.ResultAction_OneNoHighScore();
-            }
-            else {
-                return Messages.ResultAction_MultipleNoHighScore(days);
-            }
-        }
-    }
-
-    private String createSuccessfulHighScoreMessage() {
-        if (result.isNewSuccessfulHighScore()) {
-            long days = getDays(result.getSuccessfulHighScore());
-            if (days == 1) {
-                return Messages.ResultAction_SuccessfulOneHighScore();
-            }
-            else {
-                return Messages.ResultAction_SuccessfulMultipleHighScore(days);
-            }
-        }
-        else {
-            long days = getDays(result.getSuccessfulHighScoreGap());
-            if (days == 1) {
-                return Messages.ResultAction_SuccessfulOneNoHighScore();
-            }
-            else {
-                return Messages.ResultAction_SuccessfulMultipleNoHighScore(days);
-            }
-        }
-    }
-
-    private String createDeltaMessage() {
-        return tool.getDeltaMessage(result.getNewSize(), result.getFixedSize());
-    }
-
-    private long getDays(final long ms) {
-        return Math.max(1, ms / DateUtils.MILLIS_PER_DAY);
-    }
-
-    */
-
     static class LabelProviderFactoryFacade {
         public StaticAnalysisLabelProvider get(final String id) {
             return new LabelProviderFactory().create(id);
