+++ src/main/resources/io/jenkins/plugins/analysis/core/model/Messages.properties
@@ -24,7 +24,7 @@ Tool.MultipleOutstandingWarnings={0} outstanding warnings
 Tool.OneFixedWarning=One fixed warning
 Tool.MultipleFixedWarnings={0} fixed warnings
 Tool.ParticipatingTools=Static analysis results from: {0}
-Tool.NoIssuesSinceBuild={0} for {1} builds (since build {2})
+Tool.NoIssuesSinceBuild={0} for {1} builds, i.e. since build {2}
 Tool.QualityGate=Quality gates:
 Tool.ReferenceBuild=Reference build {0}
